requirements.txt
opencv-python-headless==4.10.0.84
numpy~=2.0.0
Flask==3.0.3
firebase-admin==6.0.1
face-recognition==1.3.0
python-dotenv==1.0.1
dlib==19.24.4
cmake==3.29.6
opencv-python~=4.10.0.84
cvzone~=1.6.1
